fix TotalHoldingCount
